Options should reflect xpdf options, not spawning options
Currently the options object does not give any control of how xpdf should parse the PDF. Instead it gives you options for how to control the spawning process which seems to be an implementation detail not relevant for the client.

I suggest using the options object to control xpdf. I decided to go with an array instead of an object for the options argument since it was the quickest fix and does not require constant updates to this module should xpdf decide to change the list of arguments. It does break backwards compatibility though.